redux added to project!
