+++ app/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import resource from './resourceReducer';
+
+const rootReducer = combineReducers({
+  resource,
+});
+
+export default rootReducer;
