+++ app/reducers/resourceReducer.js
@@ -1,0 +1,12 @@
+import * as types from '../actions/actionTypes';
+
+const initialState = {
+  resource: null,
+};
+
+export default function resourceReducer(state = initialState, action) {
+  switch (action.type) {
+    default:
+      return state;
+  }
+}
