+++ package.json
@@ -22,7 +22,11 @@
     "react": "^15.1.0",
     "react-dom": "^15.1.0",
     "react-link": "^1.0.3",
+    "react-redux": "^5.0.1",
     "react-router": "^2.4.0",
+    "redux": "^3.6.0",
+    "redux-immutable-state-invariant": "^1.2.4",
+    "redux-thunk": "^2.1.0",
     "webpack": "^1.13.0",
     "whatwg-fetch": "^2.0.1"
   },
